Stop energy calcs from overflowing by switching int->long in several places in the new graph code to be consistent with how they're used elsewhere.

+++ src/main/java/gregtech/api/graphs/consumers/ConsumerNode.java
@@ -17,7 +17,7 @@ public class ConsumerNode extends Node {
         return !mTileEntity.isInvalid();
     }
 
-    public int injectEnergy(int aVoltage, int aMaxAmps) {
+    public int injectEnergy(long aVoltage, long aMaxAmps) {
         return 0;
     }
 }
