+++ src/main/java/gregtech/api/graphs/consumers/NodeEnergySink.java
@@ -18,7 +18,7 @@ public class NodeEnergySink extends ConsumerNode {
     }
 
     @Override
-    public int injectEnergy(int aVoltage, int aMaxAmps) {
+    public int injectEnergy(long aVoltage, long aMaxAmps) {
         int tUsedAmps = 0;
         while (aMaxAmps > tUsedAmps && ((IEnergySink) mTileEntity).getDemandedEnergy() > 0 &&
                 ((IEnergySink) mTileEntity).injectEnergy(ForgeDirection.getOrientation(mSide), aVoltage, aVoltage) < aVoltage)
