+++ src/main/java/gregtech/api/graphs/paths/PowerNodePath.java
@@ -5,14 +5,14 @@ import gregtech.api.metatileentity.MetaPipeEntity;
 import gregtech.api.metatileentity.implementations.GT_MetaPipeEntity_Cable;
 import net.minecraft.server.MinecraftServer;
 
-//path for cables
-//all calculations like amp and voltage happens here
+// path for cables
+// all calculations like amp and voltage happens here
 public class PowerNodePath extends NodePath {
-    int mMaxAmps;
-    int mAmps = 0;
-    int mLoss;
-    int mVoltage = 0;
-    int mMaxVoltage;
+    long mMaxAmps;
+    long mAmps = 0;
+    long mLoss;
+    long mVoltage = 0;
+    long mMaxVoltage;
     int mTick = 0;
     boolean mCountUp = true;
 
@@ -21,11 +21,11 @@ public class PowerNodePath extends NodePath {
         super(aCables);
     }
 
-    public int getLoss() {
+    public long getLoss() {
         return mLoss;
     }
 
-    public void applyVoltage(int aVoltage, boolean aCountUp) {
+    public void applyVoltage(long aVoltage, boolean aCountUp) {
         int tNewTime = MinecraftServer.getServer().getTickCounter();
         if (mTick != tNewTime) {
             reset(tNewTime - mTick);
@@ -55,7 +55,7 @@ public class PowerNodePath extends NodePath {
         mAmps = Math.max(0, mAmps - (mMaxAmps * aTimePassed));
     }
 
-    public void addAmps(int aAmps) {
+    public void addAmps(long aAmps) {
         this.mAmps += aAmps;
         if (this.mAmps > mMaxAmps * 40) {
             for (MetaPipeEntity tCable : mPipes) {
@@ -69,9 +69,9 @@ public class PowerNodePath extends NodePath {
         }
     }
 
-    //if no amps pass trough for more then 0.5 second reduce them to minimize wrong results
-    //but still allow the player to see if activity is happening
-    public int getAmps() {
+    // if no amps pass through for more than 0.5 second reduce them to minimize wrong results
+    // but still allow the player to see if activity is happening
+    public long getAmps() {
         int tTime = MinecraftServer.getServer().getTickCounter() - 10;
         if (mTick < tTime) {
             reset(tTime - mTick);
@@ -80,7 +80,7 @@ public class PowerNodePath extends NodePath {
         return mAmps;
     }
 
-    public int getVoltage(MetaPipeEntity aCable) {
+    public long getVoltage(MetaPipeEntity aCable) {
         int tLoss = 0;
         if (mCountUp) {
             for (int i = 0; i < mPipes.length; i++) {
@@ -109,9 +109,9 @@ public class PowerNodePath extends NodePath {
         mMaxVoltage = Integer.MAX_VALUE;
         for (MetaPipeEntity tCable : mPipes) {
             if (tCable instanceof GT_MetaPipeEntity_Cable) {
-                mMaxAmps = Math.min((int) ((GT_MetaPipeEntity_Cable) tCable).mAmperage, mMaxAmps);
-                mLoss += (int) ((GT_MetaPipeEntity_Cable) tCable).mCableLossPerMeter;
-                mMaxVoltage = Math.min((int) ((GT_MetaPipeEntity_Cable) tCable).mVoltage, mMaxVoltage);
+                mMaxAmps = Math.min(((GT_MetaPipeEntity_Cable) tCable).mAmperage, mMaxAmps);
+                mLoss += ((GT_MetaPipeEntity_Cable) tCable).mCableLossPerMeter;
+                mMaxVoltage = Math.min(((GT_MetaPipeEntity_Cable) tCable).mVoltage, mMaxVoltage);
             }
         }
     }
